✨ (fetching): Created helper fns for CMS com
Created two functions, one for getting the enrollment of a student based on their user ID and the other to get the mission and it's components based on slug. Added isVideo function to determine whether a checkpoint is type quiz or video, returns true for video and false for quiz.